Use cast in fprintf for val.i.

+++ mir-tests/run-test.c
@@ -55,7 +55,7 @@ int main (int argc, char *argv[]) {
   if (interpr_p) {
     MIR_link (ctx, MIR_set_interp_interface, NULL);
     MIR_interp (ctx, main_func, &val, 0);
-    fprintf (stderr, "%s: %lu\n", mir_fname, val.i);
+    fprintf (stderr, "%s: %lu\n", mir_fname, (unsigned long) val.i);
   } else if (gen_p) {
     MIR_gen_init (ctx);
 #if MIR_GEN_DEBUG
